+++ airbnb-app/src/App.js
@@ -2,7 +2,7 @@ import { CustomThemeProvider } from './custom-styled-component/CustomThemeProvid
 import { ThemeProvider } from 'styled-components';
 import { GlobalStyle } from './common/globalStyle';
 import theme from './common/theme';
-import GNB from './component/GNB';
+import GNB from './component/gnb/GNB';
 import SearchBar from './component/search-bar/SearchBar';
 
 function App() {
