+++ airbnb-app/src/component/gnb/Logo.jsx
@@ -1,11 +1,10 @@
-import styled from 'styled-components';
+import customStyled from '../../custom-styled-component/customStyled';
 
 function Logo() {
   return <StyledTitle>LOGO</StyledTitle>;
 }
 
-//const StyledTypography = styled(Typography)``
-const StyledTitle = styled.h1`
+const StyledTitle = customStyled.h1`
   width: 88px;
   height: 46px;
   line-height: 46px;
